adjust pricefeed.sol according to pyth updates

+++ src/contracts/core/interfaces/IPriceFeed.sol
@@ -3,7 +3,7 @@
 pragma solidity 0.8.19;
 
 interface IPriceFeed {
-    function getPriceOfToken(address _token) external view returns(uint);
-    function getMaxPriceOfToken(address _token) external view returns(uint);
-    function getMinPriceOfToken(address _token) external view returns(uint);
+    function getPriceOfToken(address _token) external view returns(uint256);
+    function getMaxPriceOfToken(address _token) external view returns(uint256);
+    function getMinPriceOfToken(address _token) external view returns(uint256);
 }
